requirements.txt
# Core Django requirements
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0

# Database
psycopg2-binary==2.9.9

# Async processing
celery==5.3.4
django-celery-results==2.5.1

# Face recognition and image processing
deepface==0.0.79
opencv-python==4.8.1.78
mtcnn==0.1.1

# Video processing
moviepy==1.0.3

# Additional utilities
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
